requirements.txt
python-telegram-bot[ext]~=22.7
telethon~=1.37
sqlalchemy~=2.0
aiosqlite~=0.20
python-dotenv~=1.0
